requirements.txt
gradio==4.44.0
pycryptodome==3.20.0